+++ demisto_sdk/__main__.py
@@ -985,11 +985,15 @@ def generate_docs(**kwargs):
     help="Output file path, the default is the Tests directory.",
     default=''
 )
+@click.option(
+    "--fail-duplicates", help="Fails the process if any duplicates are found.", is_flag=True
+)
 def create_id_set(**kwargs):
     """Create the content dependency tree by ids."""
     check_configuration_file('create-id-set', kwargs)
     id_set_creator = IDSetCreator(**kwargs)
-    id_set_creator.create_id_set(fail_on_duplicates=True)
+    fail_on_duplicates = kwargs.get('fail_duplicates')
+    id_set_creator.create_id_set(fail_on_duplicates)
 
 
 # ====================== merge-id-sets ====================== #